+++ features/admin.authentication-flow-builder-core.v1/components/decorated-visual-flow.tsx
@@ -20,8 +20,8 @@ import { IdentifiableComponentInterface } from "@wso2is/core/models";
 import { ReactFlowProvider } from "@xyflow/react";
 import classNames from "classnames";
 import React, { FunctionComponent, HTMLAttributes, ReactElement } from "react";
-import ElementPropertiesPanel from "./element-properties-panel";
-import ElementsPanel from "./elements-panel";
+import ElementPropertiesPanel from "./element-properties/element-properties-panel";
+import ElementsPanel from "./element-properties/elements-panel";
 import VisualFlow from "./visual-flow";
 import useAuthenticationFlowBuilderCore from "../hooks/use-authentication-flow-builder-core-context";
 import DnDProvider from "../providers/dnd-provider";
