+++ features/admin.authentication-flow-builder-core.v1/components/element-panel/element-panel-draggable-node.scss
