+++ features/admin.authentication-flow-builder-core.v1/components/element-panel/elements-panel.scss
