+++ features/admin.authentication-flow-builder-core.v1/components/element-properties/element-properties-panel.scss
